refactor(tests): split admin pressure analysis coverage
Move the pressure analysis regression into its own server test module.

Keep the legacy admin users suite under the rust source line budget.
